pyproject.toml
[tool.poetry]
name = "summarization"
version = "0.1.0"
description = ""
authors = ["Martin Roed <msroed@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
databutton2 = "^0.7.9"
transformers = "^4.19.2"
tokenizers = "0.11.6"
torch = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
